Modularize hero social hook
